added sidebar nav styling
