+++ frontend/src/index.js
@@ -20,7 +20,6 @@ import CreateRocket from "./components/CreateRocket.js";
 import { Elements, StripeProvider } from "react-stripe-elements";
 import BillingForm from "./components/Billing";
 import GetUser from "./components/GetUser.js";
-import Home from "./components/Home.js";
 
 const composeEnhancer = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
 
@@ -53,7 +52,6 @@ ReactDOM.render(
             />
             {/* <Route path="/billing" component={BillingForm fontSize={elementFontSize}} /> */}
           </Elements>
-          <Route path="/home" component={Home} />
         </div>
       </Router>
     </StripeProvider>
